# docs/swagger.yaml
openapi: 3.0.3
info:
  title: FanFare API
  description: |
    Pop Culture Prediction & Sentiment Platform Backend API.
    
    FanFare enables fans to predict outcomes, track community sentiment,
    contribute crowdsourced context cards, and build reputation through
    accurate insights — not follower counts.
    
    ## 4 Key Product Pillars
    1. **Predictions Drive Engagement** — Users predict box office, awards, ratings, story events
    2. **Sentiment Dashboards Show the Fandom Pulse** — Track hype, satisfaction, controversies over time
    3. **Context Cards Reduce Misinformation** — Structured crowd-sourced notes clarify truth vs rumor
    4. **Reputation from Insight, Not Followers** — Points for accurate predictions & helpful contributions
  version: 1.0.0
  contact:
    name: FanFare Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    Universe:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [movie, tv, mixed]
        primary_languages:
          type: array
          items:
            type: string
        region_tags:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
    Title:
      type: object
      properties:
        id:
          type: string
        universe_id:
          type: string
        name:
          type: string
        slug:
          type: string
        release_date:
          type: string
        synopsis:
          type: string
        poster_art_url:
          type: string
        cast_list:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [announced, in_production, released, ended]
    Moment:
      type: object
      properties:
        id:
          type: string
        universe_id:
          type: string
        title_id:
          type: string
        title:
          type: string
        summary:
          type: string
        type:
          type: string
          enum: [official_news, industry_move, rumor_leak, fan_theory]
        source_urls:
          type: array
          items:
            type: string
        source_type:
          type: string
          enum: [studio, trade, gossip, user]
    IssueHub:
      type: object
      properties:
        id:
          type: string
        universe_id:
          type: string
        title:
          type: string
        slug:
          type: string
        overview:
          type: string
        timeline:
          type: array
          items:
            type: object
            properties:
              moment_id:
                type: string
              event_text:
                type: string
              date:
                type: string
        evidence_stack:
          type: array
          items:
            type: object
            properties:
              metric_name:
                type: string
              value:
                type: string
              type:
                type: string
    Prediction:
      type: object
      properties:
        id:
          type: string
        moment_id:
          type: string
        issue_hub_id:
          type: string
        question:
          type: string
        type:
          type: string
          enum: [binary, threshold]
        options:
          type: array
          items:
            type: string
        resolution_criteria:
          type: string
        status:
          type: string
          enum: [open, resolved, cancelled]
        correct_option:
          type: string
        creator_id:
          type: string
...
    ContextCardRequest:
      type: object
      required: [whats_accurate, whats_missing, whats_speculative]
      properties:
        moment_id:
          type: string
        issue_hub_id:
          type: string
        prediction_id:
          type: string
        whats_accurate:
          type: string
          minLength: 10
        whats_missing:
          type: string
          minLength: 10
        whats_speculative:
          type: string
          minLength: 10
    ContextCard:
      type: object
      properties:
        id:
          type: string
        moment_id:
          type: string
        issue_hub_id:
          type: string
        prediction_id:
          type: string
        user_id:
          type: string
        whats_accurate:
          type: string
        whats_missing:
          type: string
        whats_speculative:
          type: string
        status:
          type: string
        helpful_count:
          type: integer
        not_helpful_count:
          type: integer
        biased_count:
          type: integer
    ContextCardRateRequest:
      type: object
      required: [rating]
      properties:
        rating:
          type: string
          enum: [helpful, not_helpful, biased]
...
paths:
...
  /api/v1/universes:
    get:
      summary: List all pop-culture universes
      tags: [Universes]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of universes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Universe'
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/universes/{slug}:
    get:
      summary: Get universe with its titles
      tags: [Universes]
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Universe details with titles
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: Universe not found
  /api/v1/titles/{slug}:
    get:
      summary: Get title by slug
      tags: [Titles]
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Title details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Title'
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: Title not found
  /api/v1/moments:
    get:
      summary: Get paginated news feed moments
      tags: [Moments]
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Feed of moments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Moment'
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/moments/{id}:
    get:
      summary: Get a specific moment
      tags: [Moments]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Moment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Moment'
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: Moment not found
  /api/v1/moments/{id}/predictions:
    get:
      summary: List all predictions attached to a moment
      description: Returns open and resolved prediction questions linked to this news card
      tags: [Moments]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prediction'
        '401':
          description: Unauthorized — valid JWT required
        '400':
          description: Invalid moment ID
  /api/v1/issues:
    get:
      summary: List debate hubs
      tags: [Issues]
      security:
        - BearerAuth: []
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, activity]
            default: recent
      responses:
        '200':
          description: List of issue hubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssueHub'
        '401':
          description: Unauthorized — valid JWT required
  /api/v1/issues/{slug}:
    get:
      summary: Get debate hub with timeline and evidence
      tags: [Issues]
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue hub details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueHub'
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: Issue hub not found
  /api/v1/issues/{slug}/predictions:
    get:
      summary: List all predictions within an issue hub debate
      tags: [Issues]
      security:
        - BearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of predictions in this debate
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prediction'
        '401':
          description: Unauthorized — valid JWT required
        '404':
          description: Issue hub not found
...
  /api/v1/context:
    post:
      summary: Submit a context card
      tags: [Context]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextCardRequest'
      responses:
        '201':
          description: Context card created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextCard'
  /api/v1/context/{id}/rate:
    post:
      summary: Rate a context card
      tags: [Context]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextCardRateRequest'
      responses:
        '200':
          description: Rating submitted
        '409':
          description: Already rated